Backend/requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==1.11.10
django-constance==2.1.0
django-cors-middleware==1.3.1
django-filter==1.1.0
django-jet==1.0.7
django-oauth-toolkit==1.0.0
django-picklefield==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
hashids==1.2.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
